Tests: Update test_common to reflect move of get_species_by_label()
The function was moved from main.py to common.py so the corresponding test was moved from test_main.py to test_common.py
