+++ tests/test_main.py
@@ -784,20 +784,6 @@ def test_load_species():
 
 # main functions:
 
-
-def test_get_species_by_label():
-    """Test getting species by label"""
-    t3 = run_minimal(project_directory=os.path.join(DATA_BASE_PATH, 'minimal_data'),
-                     iteration=1,
-                     set_paths=True,
-                     )
-    rmg_species, rmg_reactions = t3.load_species_and_reactions_from_chemkin_file()
-    label = 'H2O'
-    species = get_species_by_label(label, rmg_species)
-    assert species.label == label
-    assert species.index == 7
-
-
 def test_get_reaction_by_index():
     """Test getting reaction by index"""
     t3 = run_minimal(project_directory=os.path.join(DATA_BASE_PATH, 'minimal_data'),
